new Login Screen UI, completion of sorting and charts in statisticsVC

+++ ExpenseIncomeTracker/Models/ModelTransactionDetails.swift
@@ -12,7 +12,7 @@ struct Transaction: Codable {
     
     let date: String
     
-    let amount: Double
+    let amount: Int
     
     let name: String
